+++ package.json
@@ -10,7 +10,7 @@
     "lint:fix": "eslint --fix",
     "test": "npm run test:unit && npm run test:typescript",
     "test:unit": "c8 --100 node --test",
-    "test:typescript": "tsd"
+    "test:typescript": "tstyche"
   },
   "repository": {
     "type": "git",
@@ -61,7 +61,7 @@
     "fastify": "^5.0.0",
     "h2url": "^0.2.0",
     "neostandard": "^0.13.0",
-    "tsd": "~0.33.0"
+    "tstyche": "^7.0.0"
   },
   "dependencies": {
     "fast-uri": "^3.0.0",
